Move tests from src/Tests to tests/src
Ensure namespace best practise is followed
Remove broken tests
Clean up SetUp and TearDown hooks
