+++ tests/src/Kernel/RabbitMqBaseTest.php
@@ -1,17 +1,18 @@
 <?php
 
-namespace Drupal\rabbitmq\Tests;
+namespace Drupal\Tests\rabbitmq\Kernel;
 
+use Drupal\KernelTests\KernelTestBase;
 use Drupal\rabbitmq\Queue\QueueBase;
 use Drupal\rabbitmq\Queue\QueueFactory;
-use Drupal\KernelTests\KernelTestBase;
+
 use PhpAmqpLib\Channel\AMQPChannel;
 use PhpAmqpLib\Connection\AMQPStreamConnection;
 
 /**
  * Class RabbitMqTestBase is a base class for RabbitMQ tests.
  */
-abstract class RabbitMqTestBase extends KernelTestBase {
+abstract class RabbitMqBaseTest extends KernelTestBase {
   const MODULE = 'rabbitmq';
 
   public static $modules = ['system', QueueBase::MODULE];
@@ -75,8 +76,25 @@ abstract class RabbitMqTestBase extends KernelTestBase {
   }
 
   /**
+   * {@inheritdoc}
+   */
+  public function tearDown() {
+    $connection = $this->connectionFactory->getConnection();
+
+    $channel = $connection->channel();
+    if ($channel instanceof AMQPChannel) {
+      $channel->close();
+    }
+
+    $connection->close();
+
+    parent::tearDown();
+  }
+
+  /**
    * Initialize a server and free channel.
    *
+   * @param string $name
    * @return array[]
    *   - \AMQPChannel: A channel to the default queue.
    *   - string: the queue name.
@@ -84,29 +102,19 @@ abstract class RabbitMqTestBase extends KernelTestBase {
   protected function initChannel($name = QueueFactory::DEFAULT_QUEUE_NAME) {
     $connection = $this->connectionFactory->getConnection();
     $this->assertTrue($connection instanceof AMQPStreamConnection, 'Default connections is an AMQP Connection');
+
     $channel = $connection->channel();
     $this->assertTrue($channel instanceof AMQPChannel, 'Default connection provides channels');
-    $passive = FALSE;
-    $durable = FALSE;
-    $exclusive = FALSE;
-    $auto_delete = TRUE;
 
-    list($actual_name,,) = $channel->queue_declare($name, $passive, $durable, $exclusive, $auto_delete);
-    $this->assertEquals($name, $actual_name, 'Queue declaration succeeded');
+    list($actualName,,) = $channel->queue_declare(
+      $name,
+      FALSE,
+      TRUE,
+      FALSE,
+      FALSE
+    );
+    $this->assertEquals($name, $actualName, 'Queue declaration succeeded');
 
-    return [$channel, $actual_name];
+    return [$channel, $actualName];
   }
-
-  /**
-   * Clean up after a test.
-   *
-   * @param \PhpAmqpLib\Channel\AMQPChannel $channel
-   *   The channel to clean up.
-   */
-  protected function cleanUp(AMQPChannel $channel) {
-    $connection = $channel->getConnection();
-    $channel->close();
-    $connection->close();
-  }
-
 }
